Refactor color handling and configuration: update Defaults module to use nested classes for themes, colors, and config, enhancing organization and clarity
